+++ server/src/test/java/org/elasticsearch/node/NodeTests.java
@@ -19,6 +19,7 @@ import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.common.transport.BoundTransportAddress;
 import org.elasticsearch.core.RestApiVersion;
 import org.elasticsearch.env.Environment;
+import org.elasticsearch.index.IndexModule;
 import org.elasticsearch.index.IndexService;
 import org.elasticsearch.index.engine.Engine.Searcher;
 import org.elasticsearch.index.shard.IndexShard;
@@ -445,6 +446,51 @@ public class NodeTests extends ESTestCase {
                 expectThrows(NamedObjectNotFoundException.class, () -> namedXContentRegistry.lookupParser(Integer.class, "name", p));
             }
         }
+    }
+
+    static class AdditionalSettingsPlugin1 extends Plugin {
+        @Override
+        public Settings additionalSettings() {
+            return Settings.builder()
+                .put("foo.bar", "1")
+                .put(IndexModule.INDEX_STORE_TYPE_SETTING.getKey(), IndexModule.Type.MMAPFS.getSettingsKey())
+                .build();
+        }
+    }
+
+    static class AdditionalSettingsPlugin2 extends Plugin {
+        @Override
+        public Settings additionalSettings() {
+            return Settings.builder().put("foo.bar", "2").build();
+        }
+    }
+
+    public void testAdditionalSettings() {
+        Settings settings = Settings.builder()
+            .put(Environment.PATH_HOME_SETTING.getKey(), createTempDir())
+            .put("my.setting", "test")
+            .put(IndexModule.INDEX_STORE_TYPE_SETTING.getKey(), IndexModule.Type.NIOFS.getSettingsKey())
+            .build();
+        Map<String, Plugin> pluginMap = Map.of(AdditionalSettingsPlugin1.class.getName(), new AdditionalSettingsPlugin1());
+        Settings newSettings = Settings.builder().put(Node.mergedPluginSettings(pluginMap)).put(settings).build();
+        assertEquals("test", newSettings.get("my.setting")); // previous settings still exist
+        assertEquals("1", newSettings.get("foo.bar")); // added setting exists
+        // does not override pre existing settings
+        assertEquals(IndexModule.Type.NIOFS.getSettingsKey(), newSettings.get(IndexModule.INDEX_STORE_TYPE_SETTING.getKey()));
+    }
+
+    public void testAdditionalSettingsClash() {
+        Map<String, Plugin> pluginMap = Map.of(
+            AdditionalSettingsPlugin1.class.getName(),
+            new AdditionalSettingsPlugin1(),
+            AdditionalSettingsPlugin2.class.getName(),
+            new AdditionalSettingsPlugin2()
+        );
+        IllegalArgumentException e = expectThrows(IllegalArgumentException.class, () -> Node.mergedPluginSettings(pluginMap));
+        String msg = e.getMessage();
+        assertTrue(msg, msg.contains("Cannot have additional setting [foo.bar]"));
+        assertTrue(msg, msg.contains("plugin [" + AdditionalSettingsPlugin1.class.getName()));
+        assertTrue(msg, msg.contains("plugin [" + AdditionalSettingsPlugin2.class.getName()));
     }
 
     private RestRequest request(NamedXContentRegistry namedXContentRegistry, RestApiVersion restApiVersion) throws IOException {
